requirements.txt
Flask==0.12.2
google-cloud-language==0.23.1
google-cloud-bigquery==0.23.0

